+++ src/test/java/techcourse/myblog/domain/repository/ArticleRepositoryTest.java
@@ -1,9 +1,9 @@
-package techcourse.myblog.domain;
+package techcourse.myblog.domain.repository;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import techcourse.myblog.domain.repository.ArticleRepository;
+import techcourse.myblog.domain.Article;
 
 import static org.assertj.core.api.Assertions.assertThat;
