+++ src/test/java/techcourse/myblog/domain/repository/UserRepositoryTest.java
@@ -1,10 +1,10 @@
-package techcourse.myblog.domain;
+package techcourse.myblog.domain.repository;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import techcourse.myblog.domain.repository.UserRepository;
+import techcourse.myblog.domain.User;
 
 import static org.assertj.core.api.Assertions.assertThat;
